+++ CHANGELOG.md
@@ -6,9 +6,11 @@ All notable changes to this project will be documented in this file.
 
 ### Changed
 
-- Replace stackable-operator `initialize_logging` with stackable-telemetry `Tracing` ([#338]).
-  - BREAKING: The file log directory was set by `COMMONS_OPERATOR_LOG_DIRECTORY`,
-    and is now set by `ROLLING_LOGS` (or via `--rolling-logs <DIRECTORY>`).
+- Replace stackable-operator `initialize_logging` with stackable-telemetry `Tracing` ([#338], [#344]).
+  - BREAKING: The console log level was set by `COMMONS_OPERATOR_LOG`, and is now set by `CONSOLE_LOG`.
+  - BREAKING: The file log level was set by `COMMONS_OPERATOR_LOG`, and is now set by `FILE_LOG`.
+  - BREAKING: The file log directory was set by `COMMONS_OPERATOR_LOG_DIRECTORY`, and is now set
+    by `ROLLING_LOGS_DIR` (or via `--rolling-logs <DIRECTORY>`).
   - Replace stackable-operator `print_startup_string` with `tracing::info!` with fields.
 
 ### Fixed
@@ -17,6 +19,7 @@ All notable changes to this project will be documented in this file.
 
 [#338]: https://github.com/stackabletech/commons-operator/pull/338
 [#343]: https://github.com/stackabletech/commons-operator/pull/343
+[#344]: https://github.com/stackabletech/commons-operator/pull/344
 
 ## [25.3.0] - 2025-03-21